move context into hook files
