--- src/components/Squeak/hooks/useQuestion.ts
@@ -1,7 +1,0 @@
-import { useContext } from 'react'
-import { Context } from '../context/question'
-
-export const useQuestion = () => {
-  const question = useContext(Context)
-  return question
-}
