--- src/components/Squeak/hooks/useUser.ts
@@ -1,7 +1,0 @@
-import { useContext } from 'react'
-import { Context } from '../context/user'
-
-export const useUser = () => {
-    const user = useContext(Context)
-    return user
-}
