+++ src/components/Squeak/index.ts
@@ -6,26 +6,24 @@ import { Squeak } from './components/main/Squeak'
 import { Days } from './components/Days'
 import { Avatar } from './components/Avatar'
 import { Login } from './components/main/Login'
-import { Provider as OrgProvider } from './context/org'
-import { Provider as UserProvider } from './context/user'
 import { Questions } from './components/main/Questions'
-import { useOrg } from './hooks/useOrg'
-import { useUser } from './hooks/useUser'
+import { useOrg, Provider as OrgProvider } from './hooks/useOrg'
+import { useUser, Provider as UserProvider } from './hooks/useUser'
 import { useQuestion } from './hooks/useQuestion'
 
 export {
-  Authentication,
-  Squeak,
-  FullQuestion,
-  Questions,
-  Question,
-  Form,
-  Login,
-  Avatar,
-  Days,
-  OrgProvider,
-  UserProvider,
-  useOrg,
-  useUser,
-  useQuestion
+    Authentication,
+    Squeak,
+    FullQuestion,
+    Questions,
+    Question,
+    Form,
+    Login,
+    Avatar,
+    Days,
+    OrgProvider,
+    UserProvider,
+    useOrg,
+    useUser,
+    useQuestion,
 }
